My First Noodle: the Yelp logo!
This is just following from the lines2.asm chrissalo example, but by
using Mov_Draw_VLc_a instead (which moves the beam to the first vector
before drawing). First move is still relative, which means that the
later 4 'sections' of the logo (9oclock, 7oclock, 4oclock, 2oclock) are
moved relative to the previous section rather than the origin 0,0 point.

I'm probably missing a trick with that, but i did try just doing a Wait_Recal
beforehand, and that didn't go well at all (scaling increased?).

Also found out that you need to be careful with the SPRITE_BLOW_UP value, as I
was overflowing something when SPRITE_BLOW_UP EQL 25.